printing_prime.py KLAAAAR!!!!

+++ 1DV501/lc222rx_assign2/distance.py
@@ -1,10 +1,10 @@
 # A program that calculates the distance between two points
 from math import sqrt
 
-x1 = int(input("Enter x1: "))
-y1 = int(input("Enter y1: "))
-x2 = int(input("Enter x2: "))
-y2 = int(input("Enter y2: "))
+x1 = float(input("Enter x1: "))
+y1 = float(input("Enter y1: "))
+x2 = float(input("Enter x2: "))
+y2 = float(input("Enter y2: "))
 
 distance = sqrt((x1-x2)**2 + (y1-y2)**2)
